src/requirements.txt
aiogram==3.21.0
psycopg2==2.9.10
python-dotenv==1.1.1
